backend/requirements.txt
livekit-agents[openai,azure,silero]>=1.0.0
python-dotenv>=0.20.0
aiohttp>=3.8.0
requests>=2.31.0
